build: Use `buildRustPackage` instead of `crate2nix`
The code of the bot rarely changes, and the final binary derivation is
rather small.  Instead of trying to be clever by splitting dependencies
into individual derivations, just use `buildRustPackage` and be done
with it.  This saves evaluation time and is easy to cache, speeding up
daily CI runs significantly.